gl_demo ==> gfx_demo

+++ samples/gfx_demo/src/gpu/ogl/simple.frag
@@ -1,9 +1,8 @@
 #ifdef GL_ES
-precision highp float;
+precision mediump float;
 #	define HALF mediump
 #	define FULL highp
 #else
 #	define HALF
 #	define FULL
 #endif
-
